requirements.txt
# LLM & Embeddings
ollama>=0.2.0

# Vector Database
chromadb>=0.5.0

# Keyword Search
rank-bm25>=0.2.2

# Agent Framework
langgraph>=0.2.0
langchain-core>=0.2.0

# Document Processing
pymupdf>=1.24.0
python-docx>=1.1.0
openpyxl>=3.1.0
pytesseract>=0.3.10
Pillow>=10.0.0

# API Backend
fastapi>=0.111.0
uvicorn>=0.30.0

# Configuration
pyyaml>=6.0.1

# Database
# sqlite3 is built into Python - no install needed

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Utilities
python-multipart>=0.0.9
aiofiles>=23.0.0
